Multiple peers connect dynamically in a mesh topology
Problems: performance, redundancy.
In order to find peers, a peer will scan an entire collection. Bad. That takes a lot of time and grows very slow as the collection expands, leaking memory as it does.

Each peer tries to initiate a connection with every other peer while listening on it's own channel for pings. There is no restriction on what  peers are greeted, meaning that when connecting to a peer, that peer can also connect back to you without knowing it, duplicating the connections and doubling memory/performance requirements.