Peer self-signaling prevention and peer GC.
On occasion, a signal would be reflected back to the peer and it would signal itself. Feeding in self-generated data will kill the peer, so new safeguards have been added to prevent it.

When a peer disconnects, now it will be removed from the collection and references will be removed.