Begin integration with gun driver handlers, automatic upgrades/peer connection from peer server settings, code cleaning, data exchange events, peer broadcast method, fix server reboot bug, switch default server port.
Now there's a handler for `gun.get` that broadcasts to the connected peers asking if they have certain data, and creates a dedicated event channel for subsequent responses.

Now webRTC creates the connection by grabbing the peers you've passed the first Gun constructor (this will probably change to create a scoped list of peers per instantiation, depending on what URLs are pointed towards).

There were a few places where the code was starting to rot. That's been cleaned up a bit, and shouldn't pose a problem for a while longer.

There's now two methods for the peer connection object, "each" and "broadcast". "each" is like a forEach loop for an array, but instead on an object. "broadcast" sends data to every connected peer.

The server was restarting each time the data.json file changed because `nodemon` was scanning the project and saw a change. Whoops. The npm start script has been changed to  use the plain `node` command instead.

Now instead of serving on port 3000 by default, it's held on 8080, yielding exponential performance gains. Okay, not really, it's just a preference.